Adding arc calculations as well.

+++ scripts/fluxRun.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-if [ $# -ne 3 ]
+if [ $# -ne 3 ] && [ $# -ne 4 ]
 then
-	echo "Usage:" $0 "nmax states contourID"
+	echo "Usage:" $0 "nmax states contourID [gval]"
 	exit 1
 fi
 
@@ -19,7 +19,11 @@ dir="nmax"$1"states"$2"contour"$3
 
 GV=`ls input/$dir | grep eigenTwo | sed -n -e 's@\(eigenTwo_\)\([-0-9e\.]*\)\(\.dat\)@\2@p'`
 
-rm -rf output/$dir
+if ! [ $# -eq 4 ]
+then
+	rm -rf output/$dir
+fi
+
 mkdir -p output/$dir
 mkdir -p output/$dir/densities
 mkdir -p output/$dir/gradients
@@ -28,6 +32,11 @@ outputName="output/$dir/res.txt"
 
 for g in $GV
 do
+	if [ "$g" != "$4" ]
+	then
+		continue
+	fi
+
 	echo "Running with g=$g"
 	tmpFile=`mktemp`
 	cp fluxConfig.conf $tmpFile
